+++ apps/payment-service/src/main/java/com/server/payment/domain/PaymentService.java
@@ -12,7 +12,7 @@ public class PaymentService implements IPaymentService {
 
     public PaymentEntity create(PaymentEntity payment) {
         Random random = new Random();
-        Long paymentId = random.nextLong(1000000); // Generates a random payment ID between 0 and 999999
+        Integer paymentId = random.nextInt(1000000); // Generates a random payment ID between 0 and 999999
         payment.setId(paymentId);
         return payment;
     }
@@ -24,7 +24,7 @@ public class PaymentService implements IPaymentService {
         int randomNumber = 3 + random.nextInt(7); // Generates a random number between 3 and 9
         for (int i = 0; i < randomNumber; i++) {
             PaymentEntity payment = new PaymentEntity();
-            payment.setId(random.nextLong(1000000)); // Generates a random payment ID between 0 and 999999
+            payment.setId(random.nextInt(1000000)); // Generates a random payment ID between 0 and 999999
             payment.setPayerEmail(payment.getId() + "@email.com");
             payment.setStatus(i % 2 == 0 ? PaymentEntity.Status.PAID : PaymentEntity.Status.PENDING);
             payment.setCurrency("EUR");
@@ -38,7 +38,7 @@ public class PaymentService implements IPaymentService {
         return payments;
     }
 
-    public PaymentEntity getById(Long id) {
+    public PaymentEntity getById(Integer id) {
         // Placeholder for implementation to retrieve a payment by its ID
         PaymentEntity payment = new PaymentEntity();
         payment.setId(id);
@@ -53,7 +53,7 @@ public class PaymentService implements IPaymentService {
         return payment;
     }
 
-    public PaymentEntity updateStatusById(Long id, Status paymentStatust) {
+    public PaymentEntity updateStatusById(Integer id, Status paymentStatust) {
         // Placeholder for implementation to update a payment
         PaymentEntity payment = new PaymentEntity();
         payment.setId(id);
@@ -65,7 +65,7 @@ public class PaymentService implements IPaymentService {
         return payment;
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Integer id) {
         // Placeholder for implementation to delete a payment by its ID
         System.out.println("PaymentEntity with ID " + id + " has been deleted.");
     }
